* Removes dependencies on manta; javax.swing, 	  gnu.java.nio and java.nio now compile with jikes;

+++ java/nio/channels/GatheringByteChannel.java
@@ -6,6 +6,6 @@ import java.io.*;
 
 public interface GatheringByteChannel extends WritableByteChannel
 {
-    public int write(ByteBuffer[] srcs, int offset, int length) throws IOException;
+    public long write(ByteBuffer[] srcs, int offset, int length) throws IOException;
     public int write(ByteBuffer[] srcs) throws IOException;
 }
