+++ java/nio/channels/ScatteringByteChannel.java
@@ -5,6 +5,6 @@ import java.io.*;
 
 public interface ScatteringByteChannel extends ReadableByteChannel
 {
-    public int write(ByteBuffer[] srcs, int offset, int length) throws IOException;
+    public long write(ByteBuffer[] srcs, int offset, int length) throws IOException;
     public int write(ByteBuffer[] srcs) throws IOException;
 }
